streamlit-ocr/requirements.txt
altair==5.5.0
appnope==0.1.4
asttokens==3.0.0
attrs==24.3.0
blinker==1.9.0
bottleneck==1.4.2
brotli==1.1.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.3.2
click==8.1.8
debugpy==1.8.11
decorator==5.2.1
exceptiongroup==1.2.2
executing==2.2.0
gitdb==4.0.7
gitpython==3.1.43
idna==3.7
importlib-metadata==8.6.1
ipykernel==6.29.5
ipython==8.36.0
jedi==0.19.2
jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2023.7.1
jupyter-client==8.6.3
jupyter-core==5.7.2
markupsafe==3.0.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numexpr==2.10.1
numpy==2.2.4
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow>=8.0.0,<11.0.0
pip==25.1
platformdirs==4.3.7
prompt-toolkit==3.0.51
protobuf==5.29.3
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==19.0.0
pydeck==0.9.1
pygments==2.19.1
pysocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2025.2
pyu2f==0.1.5
pyzmq==26.0.2
referencing==0.35.1
requests==2.32.3
rich==13.8.0
rpds-py==0.20.0
scipy==1.13.1
setuptools==75.1.0
six==1.16.0
smmap>=3.0.1,<5.0.0
stack-data==0.6.3
streamlit==1.38.0
tenacity==8.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
traitlets==5.14.3
typing-extensions==4.12.2
tzdata==2025.2
tzlocal==5.2
urllib3==2.2.2
validators==0.25.0
watchdog==4.0.1
wcwidth==0.2.13
zipp>=3.20.0